requirements.txt
Django==5.1.7
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
stripe==13.0.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
requests==2.32.3
requests-oauthlib==2.0.0
gunicorn==23.0.0
google-auth
google-auth-oauthlib
google-auth-httplib2
django-cors-headers==4.3.1

